Refatoração Conta Corrente

+++ Conta_Corrente.ConsoleApp/Conta_Corrente.ConsoleApp/Movimentacao.cs
@@ -8,7 +8,7 @@ namespace Conta_Corrente.ConsoleApp
 {
     public class Movimentacao
     {
-        public decimal valor;
+        public double valor;
         public TipoMovimentacao tipo;
         public string descricao;
